Move FakeTextBox to Reko.UnitTests.Mocks
